+++ lib/govuk_message_queue_consumer/consumer.rb
@@ -1,14 +1,5 @@
 module GovukMessageQueueConsumer
   class Consumer
-    # Only fetch one message at a time on the channel.
-    #
-    # By default, queues will grab messages eagerly, which reduces latency.
-    # However, that also means that if multiple workers are running one worker
-    # can starve another of work.  We're not expecting a high throughput on this
-    # queue, and a small bit of latency isn't a problem, so we fetch one at a
-    # time to share the work evenly.
-    NUMBER_OF_MESSAGES_TO_PREFETCH = 1
-
     def self.default_connection_from_env
       # https://github.com/ruby-amqp/bunny/blob/066496d/docs/guides/connecting.md#paas-environments
       if !ENV["RABBITMQ_URL"].to_s.empty?
@@ -21,17 +12,25 @@ module GovukMessageQueueConsumer
     # Create a new consumer
     #
     # @param queue_name [String] Your queue name. This is specific to your application,
-    #                            and should already exist and have a binding via puppet
+    #                            and should already exist and have a binding
+    #                            configured via Terraform.
     # @param processor [Object] An object that responds to `process`
     # @param rabbitmq_connection [Object] A Bunny connection object derived from `Bunny.new`
     # @param statsd_client [Statsd] An instance of the Statsd class
     # @param logger [Object] A Logger object for emitting errors (to stderr by default)
-    def initialize(queue_name:, processor:, rabbitmq_connection: Consumer.default_connection_from_env, statsd_client: NullStatsd.new, logger: Logger.new($stderr))
+    # @param worker_threads [Number] Size of the worker thread pool. Defaults to 1.
+    # @param prefetch [Number] Maximum number of unacked messages to allow on
+    #                          the channel. See
+    #                          https://www.rabbitmq.com/docs/consumer-prefetch
+    #                          Defaults to 1.
+    def initialize(queue_name:, processor:, rabbitmq_connection: Consumer.default_connection_from_env, statsd_client: NullStatsd.new, logger: Logger.new($stderr), worker_threads: 1, prefetch: 1)
       @queue_name = queue_name
       @processor = processor
       @rabbitmq_connection = rabbitmq_connection
       @statsd_client = statsd_client
       @logger = logger
+      @worker_threads = worker_threads
+      @prefetch = prefetch
     end
 
     def run(subscribe_opts: {})
@@ -75,13 +74,13 @@ module GovukMessageQueueConsumer
 
     def queue
       @queue ||= begin
-        channel.prefetch(self.class::NUMBER_OF_MESSAGES_TO_PREFETCH)
+        channel.prefetch(@prefetch)
         channel.queue(@queue_name, no_declare: true)
       end
     end
 
     def channel
-      @channel ||= @rabbitmq_connection.create_channel
+      @channel ||= @rabbitmq_connection.create_channel(nil, @worker_threads)
     end
   end
 end